ml-worker/requirements.txt
onnxruntime==1.9.0
requests
pillow
numpy